Pitch is now persistent in a session, and in CSV playlists
Within a session, changes to pitch for individual songs are persistent now.  The pitch settings are now saved (and loaded) with a CSV playlist.  This is a precursor to adding per-song tempo, loop points, and Dan's last-played timestamps.  The pitches are kept in a hidden column, that we might want to make visible later on, say, with a disclosure triangle or similar.