Add per-song tempo, persistent in session, saved in CSV playlist
Tempo now works like Pitch.  Changes to a song persist when the song is re-double-clicked.  And, if the # field is set, the tempo will also be saved in the CSV file.  "-1" in the tempo column means "go ahead and auto-detect".  This value will be overridden, if changed via the Tempo slider.  There's an internal (secret) boolean to turn on viewing of the tempo/pitch columns.  Might be a user preference later.  Temporarily (until release), let's show them.