Made pitch/tempo view into an Experimental feature
Also, fixed a bug in showing/hiding the Timer tab, which was making Call tabs go away when Preferences was opened and the Timers tab checkbox not changed and OK clicked.  Also tempo in Playlist is now "0" (rather than "-1") when tempo is unknown, to be consistent with pitch ==0 meaning "don't change it".