Finally fixed min window size bug on Windows
The fonts in the basicLabel*, etc. were sized too big on Windows, because Windows/Mac font sizes are not portable in Qt.  Windows fonts are much bigger, relatively speaking.  Fixing the font sizes of the often-on-a-hidden-tab Callerlab definitions now allows the overall windows size on Windows to be smaller.  I also set the initial window size on all screens to be a reasonable size, relative to the user's screen.