Added Vu Meter (woo hoo!)
Simple mono VU meter, to monitor levels.  It factors in the volume control, so it's basically a monitor of what your laptop would put out, if the laptop's volume control were to be turned up.  Helps to debug "no audio" problems, and it's fun to watch.  We might need a preference to turn it off at some point for those who are distracted.