Yes, I made the current tab mistake AGAIN
When the current tab in Designer is not set to Music Player, then the Timer Tab will not appear when you enter the Preferences dialog, select Enable Timer Tab, and click OK.  This is just dumb, but so is my mistake to not set it back.  This is not the first time I have done this.