Added session segment coloring to the clock
I think this is a cool feature.  The edges of the clock are colored with the same color used in the songTable.  There's one segment per minute, and the last type (e.g. patter, singing, etc.) wins that minute.  Only the last hour of minute colors is saved.  Doesn't yet clear out as the minute hand goes around (bug).  It's an experimental feature, default off, so an appropriate checkbox in preferences is there.