Fixed missing colorClockHelpLabel
This label was off the bottom of the music tab in the Y direction in Preferences.  It should be on the Experimental tab, as help for that feature.  I made various other small tweaks to make things line up more nicely (e.g. "Experimental" was cut off on the left, Choose button was slightly small).  Tested on Mac OSX and Windows.