Enable Lyrics tab checkbox is back
Renamed tab to "Lyrics" (Dan, you'll have a one-time re-check of the box in the Experimental tab).  Implemented crude once-per-second scrolling, but I don't have a Preference for it yet, so it is compile-time disabled right now (but it kinda works!).  Note that it only matches in the same directory as MP3 right now.  Note also that it will use .css stylesheets (I'm using a modified cuesheet2.css that makes it look better).