Feature: Guest mode (Mac OS only right now)
If you start SDP and then insert a flash drive containing a directory called "squareDanceMusic" with the same folder structure as a normal musicRootDirectory, those songs will be merged into the songTable.  Their types will have a suffix "*" to indicate "guest song" (you can filter on "*" in the Type field, too).  Later on, we can make a Preference for merge vs replace, to allow other directory names, and to make it work on Linux (straightforward) and Windows (I'm not so sure).