Feature: Automatic Patter Timer
Old: LONG TIP label shows up (in red) above the clock after N minutes of Patter (N set in Prefs).  If you hit Stop or Pause to do some teaching, the LONG TIP timer resets (not good).
New: a second-accurate patter timer automatically shows up above the clock at the start of patter.  Short (<60 sec) pauses/stops are counted as ongoing patter.  Switching to singing call/vocals/extras, or not playing patter for 60 seconds will make the patter timer go away. And, clicking on the patter timer display (e.g. "01:23"), will reset the timer to zero.  This addresses the problems that Dave had with the timer being reset, and it prepares the way for Clock Coloring to be second accurate (rather than minute accurate as it is now).