Feature: Automatic Break Timer
A "zero touch" break timer is a natural complement to the "zero touch" patter timer.  When enabled, auto-detects a break of longer than N (currently 30) seconds, and starts counting down the break.  When the break is over, flashes an END BREAK message, along with a (now negative) time remaining.  Also, reduced the amount CPU used by about half, because updating the warningLabel in AnalogClock::paintevent() was causing a repaint loop, with ~48 repaints per second.  This fix also nicely synchronizes the Patter/Break timer display with the Analog Clock display.  I think it's close to usable now, and so far I'm pretty happy with the notion of "zero touch".