Feature: Flash calls every N seconds (MAC OS X ONLY)
When enabled, flashes calls of particular level(s) in the NowPlaying label in red/italic, every N seconds.  This is similar to the SqMP3-Flash program.  Right now, N is hard coded to be 10, and the level is hard coded to "plus" only.  Preferences to control this will come later.