Feature: voice recognition for Mainstream calls
Initial checkin of voice recognition (Mac OS X only right now) for sdApp, which is the testbed/prototype for sd integration.  PocketSphinx installation is not included right now, you have to do that on your own, and then the executable name is hard-coded in sdApp.  Must also move the .dic and .jsgf files to ~/Documents/sdApp manually.  Way cool and super fast in a quiet environment!  (Too many False Positives in a noisy environment right now.) This is an extra-cost feature that Callerama has, that I always thought would be useful for sequence writing.  ASR is PocketSphinx, NLU is QRegExp.