"Explode and" now works, even with "and Roll"
This was a little tricky, because sd requires the brackets to be exactly right, before it will accept it.  For example: "[explode and [right and left thru]] and roll".   Also now allows for manual speaking of left/right brackets.  Also changed "Cancel [that]" to "Erase [that]" while I was there.