Mac OS X build is now self-contained
Mac bundle now includes a modified copy of pocketsphinx and its dependencies.  These are put there by the postBuildStepMac, which is only executed once (it modifies the executable and dylibs).  It sticks them into the right spot in the bundle (eventually), and it copies in the Hidden Markov Model files needed for the EN-US acoustic model.
The resulting bundle is self-contained and redistributable without needing a separate install of CMU Sphinx.  The .dic and .jsgf files are still copied into the bundle at build time, because these are still changing (the pocketsphinx and libs are not changing at all).