PocketSphinx binaries and build instructions now checked in
PocketSphinx is not trivial to compile for MacOSX.  To create a binary that will run on Yosemite, El Capitan, and Sierra, I built it on a Yosemite box, made it portable by mucking with the internal library paths, and copied to a flash drive.  I brought that back to my main machine and checked in the main executable, the dependent dylibs, and the  en-us (US English) acoustic models.  The .pro build steps are also modified now to copy the ps executable + dylibs + models into the MacOSX bundle.  Bundle was tested on Sierra and Yosemite.  Changes to the JSGF language model and the pronunciation dictionary should be auto-copied to the right spots, too.  Some other time we can work on moving pocketsphinx in as a dependent subproject (like taglib is now).