SD is now functional on Win32
The problem was the same that I had on Mac OS X -- the output was buffered.  Disabling this made everything work on Win32.  Same changes as for MacOS X, but in the wincon file.  Also fixed Win32 font size issue, got rid of a minor warning if PocketSphinx is not compiled, and disabled the Enable Voice Input menu item on all but Mac OS X..