ESC is now the Ctrl-U of SD
Ctrl-U is swallowed on Win32, for some reason, and we never get the QKeyEvent for the U key.  ESC is not swallowed, but the Ctrl-U sent to SD is still not recognized, so there's still a bug here.  In any case, TAB for auto-completion, ESC for erase current line are now the preferred keyboard shortcuts.