Fixed crashing bug in GlobalEventFilter
When the Preferences dialog is up, the pointer to the activeWindow is NOT a (MainWindow *).  This was causing intermittent crashes on some machines.  I missed this in my testing!  Fix is to return early if the dynamic_cast to MainWindow * fails, and just pass the key on to whoever needs it (in this case, the Preferences Dialog).