Temporary fixes to make the min window size smaller
Min window size will be smaller on all platforms, but especially on Linux.  Made Callerlab list font height smaller, and also allowed RenderArea to be smaller.  This is not a complete fix -- we need to revisit the height of the Callerlab font later.  This should tide us over for now!