Handle non-square renderareas better
It's hard in Qt to force a widget to be square.  So, instead of attempting that, I changed the renderArea so that the dance field is (essentially) infinite.  As the window size changes, the visible part of the renderArea will change, and the font size of the formation text will change (slightly), but the dancers will always stay the same size.  I think this is a good temporary solution, until we want to tackle the big custom layout effort required to keep the renderArea shaped like a square (and the effects that has on other widgets in the sd tab).