Created a small file to test JSGF weights
There's a strong tendency for PS to choose "Ends Square Thru" when I say "Heads Square Thru".  Not only is "Ends" weighted equally with "Heads", but "Ends Square Thru" is rarely legal at all.  In the short term, we can play with weights (carefully).  The ones in small.jsgf seem to help considerably, but I haven't tested in a broader context of "plus.jsgf" yet.
In the longer term, it may be better to either a) hard-code the possibilities, eliminating the combinations of calls that really never occur (like "Ends Square Thru") by enumeration, or b) put in a real NLU.