Checkin in the staging script for Win32 binaries
This script copies what it needs to the deployment directory (C:/cygwin/home/mpogue/deploySquareDesk).  A final test can be done there, and then the setup1.iss file is run to create the installer.  After the installer is run, the c:\Program Files\SDP version can again be tested.  If those tests all pass, it's generally ready to release on Win32.