Added menu item to toggle Auto-scroll of lyrics
Works well when the cuesheet2.css file is reasonable, not as well when the cuesheet makes the lyrics font too big.  But, this should work OK for now.  Requested by Bryce.  Setting is persistent.  Ideally, in the future, it would be good to make this a per-song setting, like tempo and pitch.