Added right-click context menu items for Playlists
Step 1 in the evolution of playlists.  Now you can right-click on a song, and you will get context-sensitive options to add that song to the playlist (at the top or bottom), move it up or down, and remove it.  Next step is to add similar items to the Playlists menu, since the right-click is very convenient, but hard-to-discover.