Added a number of [LEFT] calls to the sd grammar
There are a number of left calls that are distinct from the right handed version in their execution, even if the end position is the same (e.g. left pass thru).  Some are very distinct (e.g. left fan the top).  Added these to the JSGF grammar, and checked that sd understands them, too.