Added a Startup Wizard
This feature is focused on bringing people into the fold from, say, SqView.  If it works OK on Linux, I will make this dialog popup BEFORE the main window, if (and only if) it hasn't been run before.  It can be run again at any time via File > Startup Wizard.  Tested on Mac and Win10.