ESC now gets you out of Search Fields and stops music
Previously, if you were in a Search Field or a Timer field, and the music was playing, it was super difficult to figure out how to get back to SPACE and DOT controlling the music.  ESC really should always work.  Now, ESC will always get you out of editing mode, and will stop the music.  If you want to use SPACE bar to start/stop music and you're in an edit field, hit ESC, then SPACE.  Or, use Cmd-DOT to stop music, which always worked.  Cmd-DOT will NOT exit the search field.  This is intentional.