Get rid of "Show Tab Bar" on Mac OS X Sierra
This is a bug in Qt, which should have disabled the Native Mac OS X Sierra tab bar for us.  This bug appears to have been triggered when we added a "View" menu, and this new menu item was automatically added (only on Mac OS X >= Sierra).
This code removes it.