Sqldriver needs to be in the right place on Win10
Changed deployIt script to copy in the right DLL's.  Changed .iss installer file to install them into the right spot.  Fixes the "...failed to start because it could not find or load the Qt platform plugin windows'..." problem.  Next time, rename the Qt directory on Windows and retest, BEFORE releasing on Windows!  My tests passed because Qt was looking in the C:\Qt directory for the DLLs, and only developers like us have that directory.