Fixed #63: Spinning beachball on startup
On Mac OS X it takes ~15sec to start up the app, if there are lots and lots (18K) cuesheet files on the lyrics directory.  This turned out to be because following all the links at load time was very costly on Mac OS X/Qt.  I got rid of that, and moved the link following to LoadMP3 time (where only ONE alias resolution is done).  findFilesRecursively() time went down from 15sec to <1sec.