#88: ESC should clear the search fields
When pressing ESC, it should PAUSE (not STOP) the music, clear the search fields, and remove focus from the search fields.  Playback can be resumed by SPACE, if desired.  Before, it would lose it's place within the song, when all we wanted was to clear the search fields and get back to a sane state.