#93: play an "earcon" when Long Tip or Break Over
Haven't had a chance to implement this until now.  The menu items were already there, just needed to make them work.  Now that #92 is fixed, it was pretty straightforward to hook up playing a sound when the long tip timer or break timer expires.  The sounds played are "long_tip.mp3" and "break_over.mp3" in the soundfx directory.