#69: MIX should have one channel constant past center point
Implemented the slightly-misleading way of doing this.  The ear hears the current Constant Power curves as constant to the L or R of center MIX, even though it's actually not.  I want to retain the Constant Power mix, while giving the user much simpler feedback.  So, to the L of center, the R brought down linearly to zero.  Similarly on the R of center.  And, the volume stays fairly constant (audible volume does the right thing with Constant Power).