#24: in search mode, >>| goes to next song in full list (even if hidden)
Lots of these little UI bugs.  This is one of those "let's be careful to get it right" bugs.  If you select a row, then hide some rows, the previous method would select hidden rows.  New method does not.  If it can't find a non-hidden row, it goes back to the row it was on (even if hidden now).