#87: True muting of PocketSphinx input (Mac OS X)
When SquareDesk is active, and the SD tab is NOT active, the system input volume will be set to zero.  When the SD tab is active, the input volume will be restored.  When the app quits, the input volume will be restored.  This truly mutes PocketSphinx, which is a Good and Safe Thing.  However, if you make phone calls or use Siri while using SquareDesk, you should uncheck the "Disable microphone input..." Preference.