#97: expand to 6 soundFX, show names in menu
This is Part 1 of #97.  Instead of 3, we now have 6 soundFX available (Cmd-1 thru 6).  The files need to be in the soundfx directory in the musicDir, and they need to be named "<1-6>.<something>.mp3".  The number designates which menu item and key it binds to, and the <something> is what shows up in the menu (for discoverability).