#97: Starter set of SFX sounds
If the soundfx directory doesn't exist in the musicDir, one is automaticaly created, and it is populated (once) with 6 SFX sounds and the break_end and long_tip sounds.  The user can replace these at any time.  Also added a menu item to Stop SoundFX, and wired up the ESC key so that it also stops SFX.  Tested on MAC OS X, but not Win32 or Linux.