#104: Fix for Clear Search is super slow now
OK, when I committed the new font size feature #45, I didn't realize that when filterMusic() was setting the hidden() property of each row (as it goes), it was causing a relayout of the table for each row.  This took between 20 and 40ms, which is small, but with 200 rows, this added up quickly.  This change sets the mode to QHeaderView::Fixed, then makes everything hidden or not, then changes to QHeaderView::resizeToContents.  Time went down to <1ms per setHidden() call.