#110: auto switch to Lyrics shouldn't happen for Patter
We were only checking to see if the auto-switch was enabled.  Now it also checks to make sure that the type is one of the singing call types.  Also fixed a bug where Loop was set to ON only if the songType was called "patter".  It should have also looked at the user settings for patter-type strings.