Experimental coding on #107 to replace Callerlab lists with a more generic method.
If no such directory is found, this creates a <musicDir>/reference/ directory,
and prepopulates it with a set of text files for the various different levels.
Otherwise it loads the text files from those levels.

Filename format is an optional leading string with '.', which is used for sort
order and is stripped. The resulting file names are then presented to the user
in a combo box in the dance program tab.

The file format is optional whitespace for sort order, then "\d(\.\w+)\,?\s+"
for a teaching order number to be presented to the user, then a string for
the call name.

Obviously, users can munge the created files, or create new ones, to their
heart's content.