#120: Showed a little love to the VU Meter
VU meter is now more consistent with what other people do for horizontally-oriented VU meters (bars are wider than they are high).  Also, put in some dim bars for the non-lit bars, so that it's more obvious that this is a VU meter when not active.  Finally, reduced the number of bars to a more aesthetically pleasing number.