#124: bind +=/- to Tempo, Cmd-+=/- to Zoom
Requested by Bob Elling, and I think he's right.  Cmd-+=/- are used by Safari, Firefox, and Chrome to control Zoom.  We should do the same.  So, Tempo is now bound to ONLY +=/-.  Cmd is not required to change tempo, and not only that, Cmd must NOT be used now for Tempo changes.  Try it out, and see if you like it!