#132 partial fix: intro/outro was messed up
Some fields of SongSetting were not initialized in the constructor, which I believe caused the intro/outro to be divided by songLength too often.  This resulted in a YELLOW songPositionSlider, because both values were way too small.  I also introduced an operator<< for SongSetting, so that we can qDebug() << setting;