#140: windows-1252 encoded cuesheets displayed properly
When cuesheets come in with windows-1252 encoding, they are not handled by QString, which expects the file to be in Unicode.  When such a file is detected (contains "windows-1252"), we convert it to Unicode.  Tested with SIR21512.  I believe this will fix all such files, but I have not tested other examples.