Initial commit of html-tidy stuff (latest)
It's setup for Mac OS X only right now.  I ran cmake on it, then make, then copied the headers and dylib into the test123 directory.  It's dynamically linked in on Mac OS X only.  Code is #ifdef'd for Mac OS X only right now, so as not to disturb other development in the lyrics editing branches.