Fixed persistence of Save Cue Sheet directory
This must be a platform-specific thing.  I changed it to explicitly use the directory name of the saved file.  Previously, the path that canonicalPath() was returning was ".", so lastCuesheetSavePath was always equal to ".", which on Mac OS X takes you inside the .app file in the /Applications directory.  Let me know if this new code works on Linux!