#153: default the song filename format to "best guess"
I think the song filename-to-cuesheet-name matcher works really well now.  Changing the default to use it.  This will not affect people who overrode the default in the past.  If the guesser is NOT used, it can cause surprising layout problems with column width (e.g. songs of the format "XXXX - YYYY - ZZZZ" end up with a very large type  field).