#135: Menu item to download lyrics files
The SqView lyrics files are now hosted in our repo.  There's a hard-coded link to it in the App, and a menu item to download the ZIP file, and unpack it , and rescan for lyrics.  File > Download Lyrics... also has a progress bar for longer downloads.  Brought in QuaZip for the unpacking ZIP file job.  This is Mac OS X only right now, until Win32 and Linux are ready.