#164: initial cut at a Clear Line Format button
I'm not sure how to do a generalized Clear Format button using the QText Edit, so this is the next best thing.  If you put the cursor on a line, and don't select any text, a Clear Line Format button will be available.  Click it, and the current line's formatting will be cleared.  Needs more testing to look at the resulting HTML output, but I think this is a start.