#13: first cut of precise loop points
The intro/outro values are used for patter for looping.  That means that looping of singing calls will use the intro/outro points (there are NOT two separate values).  Fixed some problems with text<->double translation.  Fixed a problem where the blue brackets didn't move properly.  When playing, the loop point is much more precise, compared to when not playing.  This is because of coarse resolution of the seekBar slider.