#173: need absolute paths for playlists
A better fix would perhaps allow both absolute paths and paths relative to the root of the Music Directory.  Then old playlists would work, as well as allowing playlists to work when the musicDir is moved or renamed.  But, this should work for now.  Playlists created with Alpha4 will not work.