#187: consolidate Save/Save As behavior into File menu
If on the Music tab, you see "Save Playlist (As...)", enabled if there is a playlist with more than 0 items.  If on the SD tab, you see "Save Sequence As...", and "Save Sequence" is visible but disabled.  If on the Lyrics/Patter tab, you see "Save Lyrics/Patter (As)...", enabled if there are Lyrics/Patter loaded.