#161: Print Playlist...
Now implemented.  Also fixed a related bug where the label of the Print menu item was prematurely changed to Print Lyrics/Patter...
NOTE: This is just SIMPLE playlist printing, it's not GEMA compliant.  For that item, which is printing the list of what was actually played (AKA the Tracking List), see Issue #81.