#137: Make the song position bar bigger, and colors more contrasted
Now that we have text that says "Opener/Break/FigureN/Closer/etc", I think the key is to make the Opener/Middle Break/Closer sections very distinct and contrasted, and the others (Figures)  much more muted.  The Closer, in particular, means "this is it -- last chance to get everybody home".  So, this change makes the height of the color bands much larger (7 pixels, up from 3), and the colors are changed to RED for Opener/Break/Closer and DarkGrey for Figures 1/3, and DarkBlue for Figures 2/4.  That keeps the Figure sections distinct, but muted relative to the RED sections.  Note that SqView just alternates Blue/DarkBlue, which I personally don't find easy to parse at a glance.  Feedback on the colors would be appreciated!