#198: fix app crash
App crashes when we go to another window and back again with open # entry.  Somehow oldFocusWidget gets corrupted.  In the short term, I commented out restoring the previous widget focus, when you click on a different app, and come back in.  It works now for all EXCEPT the songTable # widgets.  If you leave one open for editing, when you come back in, the Type field will get focus, which isn't expected, but isn't horrible.