#23: Fixed Remembered Pitch is sometimes wrong
I think there is a case where the slider is set properly, but cBass.setPitch() is not called.  That's because valueChanged is not called automatically, when the value set is the same as the previous value.   I have added a manual call to on_pitchSlider_valueChanged() that will immediately call cBass.setPitch().  That should prevent this from happening, at least for this one possible failure mode.