Win32 is SUPER SLOW to load a singing call
It turns out that canonicalPath() is a very costly operation on Windows, and it was taking up to 2ms per cuesheet filename.  Multiply that by the 18682 entries in the pathStack, and it was taking more than 30 seconds to load a song.  And, it turns out we weren't using the results of that canonicalPath() call anymore, since Dan added the cuesheet fuzzy matching.  Commented it out, and load time went from ~30s to ~2 seconds on Win10 in Parallels.  Moderate improvement on Mac OS X (load time is now < 0.5s on Mac OS X).