Tweakable UI font sizes for Win32
Should not mess anything up on Linux.  Works on Mac OS X and Win32.  All the font and button sizes are tweakable now, based on the current Zoom setting.  The current Zoom setting is persisted, like before, but it's always Mac-specific now, and converted to the platform font/button sizes on-the-fly.