Fixed C2876 compile error (Windows), fixed volume slider
On Windows, the using statement generates a C2876 error, because "All overloaded forms of a function in a base class must be accessible to the derived class".  But, the using statement is actually not needed, because we already can access the members that we need in this derived class.  Also made the Volume slider into a click-to-move slider, like all the others.  Nice!