#205: long tip and break timers now allow selection of soundFX
In the spirit of tight integration between all the pieces, the long tip and break timers now can play sound FX when they expire.  At the recent Caller's Workshop, people enjoyed having a sound that told them when it was time to wrap up their slot, and it saved me from having to dance and watch the stopwatch at the same time.  This feature will allow the use of additional fun sounds!