#197: songs with no matching cuesheets can't see new one
OK, the problem turned out to be two-fold:  cuesheet2.css wasn't being loaded in for templates, and the default save directory was garbage.  So, we were accidentally trying to save the wrong text into the wrong place.  Sheesh.  Now cuesheet2.css is always loaded in at program start, and never agin.  And, the default save path is now <last place you saved cuesheets> or if there wasn't one, <musicDir>/lyrics.