#202: Enable Return/Enter and Up/Down to have special meaning in search
If you're in one of the search boxes, it will now select the first song in the filtered list.  If you hit Return/Enter, it will load that selected song.  You can use Up/Down keys to move the selection, and then hit Enter when you're ready to load.  Focus is cleared, so you can hit Space to start playing.