#219: adding songs while app is open
There's now a file watcher that watches all of the song-containing directories.  When a new file is added, or an existing file is deleted, the songTable is reloaded.  Right now, the sort order is NOT restored, and it goes back to the initial sort order.  We can fix that later.  This feature is used mainly when buying new music, and when you drop the music into the right musicDir subdirectory, there is now no longer a need to restart the app to get it to appear.