#54: warn when database might be shared
Added a simple lock into ".squaredesk/lock.txt" that contains the name of the machine holding the lock. If that file exists at startup, and the name of the machine holding the lock is different from this machine, then a warning dialog pops up.  You can "continue anyway", in which case the lock is taken over by the new machine, or you can "quit", and the lock remains with the first machine.  If the file exists at startup, and the name of the machine is OUR machine name, then we probably crashed holding the lock.  In this case, no dialog box is popped up, and we just resume holding the lock ourselves.