#231: Better default loop points
And #236 (silence detection).  The start loop point now defaults to 5% past the detected start of music. The end loop point now defaults to that, plus as many 32-beat phrases as can fit before the detected end of music.  This seems to work remarkably well on a large number of songs.  It's not QUITE seamless, because the BPM detector is not perfect, but it's pretty darn close, and no worse than the fixed points we had before.  Also fixed round-tripping of Intro/Outro points.