Got the basics of the New Lyrics editor working
The type is now keyed off the FG/BG COLORS of the text, which is unique for each character type.  I now have the stateful buttons working like the Bold/Italic buttons, which is what you'd expect.  If you stick the cursor into a HEADER section, the HEADER button will highlight, etc.  Char types are all mutually exclusive.  Bold/Italic can be applied to anything.  Currently no way to clear the char types, other than to stick the cursor into the line, and click Clear Line Format.  While this works, and it's what we had before, it's still ugly.  No output back to the file yet.