A little written doc on a tricky function
How to decode the type from FG/BG colors.  It could also be done from the font family name, if they all differ in the cuesheet2.css file.  QTextEdit is kinda limited in what you can do, so yes, this is a hack.  But, I think it will be robust for a while.