Factor out getResourceFile
Now just a function.  Note that resources are in different physical locations on different platforms, handled in the function.  Things loaded: lyrics/patter templates, internal cuesheet2.css.  Internal cuesheet is now the only one used.  If you don't like that one, you can change it, but colors are tied to the editor.