#251: crash when hotkey is pressed on some laptops
On Regina's laptop, this was the first time she was running the 0.8.3alpha1 version of SquareDesk, and she did not have any Reference tab items.  So, in fact, webview[0] was NULL, and the first attempt to access it (whenever a hotkey is pressed), resulted in a crash.
I added a new function "someWebViewHasFocus()", which checks only the webview[i] that are valid, based on the numWebviews counter.