Added "Recent" column
It looks like everything works now.  A new (optional) column called "Recent" now shows a red triangle if the song was played since the last time you did Session > Clear Recents.  I think it will be session-aware, because it just uses the age column (now float, but displayed as int), but I have not tested that.