Initial checkin of settings for BetterTouchTool (Mac OS X)
BetterTouchTool is an awesome little utility, that allows you to setup custom actions for practically everything on Mac OS X.  These settings set up the Touch Bar to have VOL +/-, TEMPO +/-, and PITCH +/- buttons.  The VOL buttons are for system volume, which is usually what we want when calling.  This means we no longer have to remember what the shortcut keys  are!