Fix crash on Win32 when SD coloring changes
Interestingly, sdpeople becomes invalid (that is not readable) on Win32, when renderArea->setCoupleColoringScheme(action->text()) is called.  I commented out that line, and no crash.  Surprisingly, all the coloring stuff still works! I also commented out all lines in that function, and everything still works (on both Mac OS X and Win32), so I'm not sure exactly what's going on, but that function doesn't seem to be needed, and it also causes crash on Win32, so I'm leaving it commented out for now, until I understand what this function is supposed to do better....