Fix Release version on Win32
We currently have copies of the Qt binaries in the windlls directory, and they get copied to the build directory at build time.  These were the 5.7 files, and they needed to be the 5.9.3 files.  Is there a better way to do this on Windows?  Release version now works, and it's quite a bit faster than the Debug version.