Dan's load song performance improvement
I can't see any downside to Dan's proposed change, after testing it on Mac OS X.  It does seem to improve song load time, so I'm committing it to Master so we can get some more experience with it.  This might help lessen the occurance of the #290 problem, until we can root cause it.