Belt and suspenders for #298
I'm still not sure what the root cause of Taminations .xml files appearing in the songTable, but RobB clearly had this happen.  Adding some code to prevent that, and also soundfx and sd files appearing there, just in case.  It turns out that this code also shaves ~10ms off the time to load the entire pathstack into the songTable.  Go figure.