#290: fix play/pause out of sync with reality
Theoretical root cause:  Stop() is called, but the playback engine doesn't stop quite yet (because it's a separate thread).  Then when Play() came along, if the stop wasn't done yet, Play was a toggle, and set the state back to kStopped.  I changed the cBass.Play() so that it is no longer a toggle.  It always plays now.  There is a new function called Pause() that should be called if playing, that will do only pause.  This should eliminate the race condition that I think is the root cause of this problem.  However, since I can't reliably reproduce the problem on my laptop, this is a fix in theory only.