Insert debugging, change to 0.9.2alpha1
This pitch problem is vexing.  Now if the pitch is found to be incorrect (checked once per second), it is fixed but it is also logged in <musicDir>/.squareDesk/debug.log .  The debug.log file is only written to by standalone apps, not if started from within QtCreator.  Since we are pretty good about not leaving in qDebug() statements, the debug.log file should stay quite small.