#321: K shortcut doesn't play correct pitch
If the next song is the same pitch as the current song, then cBass doesn't get called.  The code to do so was commented out (by me), and that was a mistake.  This was hard to reproduce, because it only happened when the current song and the next song had the same non-zero pitches, AND the K shortcut was used to load the next song.