#290: Fix for play/pause got out of sync with reality
OK, I think this one is fixed now.  I got rid of ALL cached versions of the libbass engine state (Stream_State and currentState), and if it's needed, we always ask the engine on-the-fly now.  Previous failures were 1/50 tries on old laptop.  Now, I can't get it to fail out of 500 tries on same laptop.  Hopefully fixed for good this time.