#334: Make MUTE mode more obvious
I've seen a couple of times that MUTE was on (vol==0) unexpectedly.  I suspect I hit the M key, but it took me a few moments to debug this, since the VU meter was bouncing around correctly.  The fact that it was MUTEd was not obvious.  This commit makes vol < 100 darkblue, vol < 50 a reddish-brown, and vol == 0 a bright red.  I think this will make it more obvious, and the user can just hit the M key to unmute.  If there is a bug here that incorrectly drops us into MUTE mode, that will be opened as a separate Issue in github.