Make it Mac OS X only right now
Most of the MainWindow code will be left uncompiled when on WIN or Linux.  The Wizard itself will be compiled everywhere, but it should be harmless.  The menu item to invoke the "Make Flash Drive Wizard" is hidden on WIN and Linux for now, until the rsync stuff is worked out on those platforms.