Previous Song now works more like other music players
If the music playhead is not at 0.0 (e.g. music is playing OR it was recently played), Previous Song will stop the music and just go back to the beginning of the song.  If the music is not playing AND we're at the beginning of the song, it will go to the previous song.  This is a lot like other players.  With the BT remote, if you're playing, hit PREV then PLAY, to start at the beginning of the song again.