Color and text now change at the same time
When hitting space bar to start/stop playback, if flash calls were enabled, the color of the text in the nowPlayingLabel would change before the text would.  This temporarily resulted in red song titles, and black flash calls.  This was minor, but annoying.  Now the color is always set immediately before the text is changed, so the text is always the right color now.