Couple of minor changes
As per my test tone, the alignment is still not quite right toward the end of the song, but this is a lot closer.  Also, changed the color so that it was visible on Mac OS X.  It's kind of a cool feature...not sure if it's worth it.  Worth playing around with some more though!