Well, it works!
I'm not sure whether I like it better than not having it, though.  It does mess up in a counter-intuitive way, when there is a strong pickup note to the downbeat of a measure.  I implemented "first negative-going zero crossing following the highest peak in a 0.45s window around the click INTRO button".  That seems to be pretty reliable, it's just that I can do ALMOST as well without it, now that I've practiced a lot while testing.