#389: handle SPACE and PERIOD specially in text boxes with focus
Restores function of #268.  I added the SPACE and PERIOD code into the special keypress handler AND it now looks up the code to run dynamically (using the new KeyAction hotness), and so it now runs the code in keyactions.h, rather than having a copy of the code in the keypress handler.  I also enhanced it to do the same for dateTimeIntro/Outro fields, which should address Regina's focus problem when coming back from another app with SDesk showing the Lyrics page.