#402: clean up warnings (part deux)
I C++11'd the casts, and I also changed most floats to doubles, to eliminate a lot of unnecessary conversions.  floats still remain in some of the DSP code for music processing, but elsewhere it's pretty much all double's now.  I also got rid of many unsigned ints, so as to avoid lots of conversions (I don't think it's likely that bugs will show up here, and there was at least one bug in there that was fixed, where a possibly negative value was being cast to unsigned int).  
So, this is general cleanup.  Tested on Mac OS X (but not Win10 or Linux yet).