#388: Save Playlist AS..., and default filename
Changed Save Playlist... to Save Playlist As..., and also the filename now defaults to the last file saved to in this session (it is not persistent across SquareDesk sessions -- if somebody wants this, please speak up!).  Also made the status message more useful, and got rid of the error status message, which wasn't really flagging a real error anyway.