#415: highlighted rows should be white-on-blue
OK, that was pretty disgusting.  If you just change the background color, it also messes up the horizontal header's font size.  Fixing that needed to be done in 4 places.  That should be factored out, but I think it's fine for now.  Highlighted rows are now white-on-blue, even if the songTable does not have focus.